backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.11.0
python-multipart==0.0.20
python-docx==1.2.0
pypdf==6.1.1
openai==1.107.2
psycopg[binary]==3.2.9
oss2==2.19.1
pytest==8.4.2
httpx==0.28.1
